ignore lines that are marked as unreachable
Sometimes source code is unreachable in reality, even though the compiler
is unable to infer that. With this patch the line can be marked as
unreachable, for example in a comment:

...
if (!soCall())
	exit(1); // unreachable, call never fails

When executing kcov as

kcov --unreachable=unreachable ...

That line is now ignored (unless it is reached), eliminating noise from the
coverage info.