Unpack struct file_data
Packing the structure brings no benefit on 32bit systems and plugs a
marginal hole of 4 bytes on 64bit architectures. It also creates a
misaligned starting point for the entries[] flex array at the end and
while this is typically not a problem for x86 it is reported by GCC
with the address-of-packed-member warning.

Closes #311