Add regex to --replace-src-path
This change allows to pass a regex to the --replace-src-path argument. This is useful, for example, when the build directory of the tested binary is a randomly created temp directory. Note that after the change the older usage of the parameter is still valid

Usage example:
kcov --replace-src-path=([^ ]*)/build_dir/src:/my_src_dir/"