configuration: fix incorrect handling of out-dir
Currently the argument parsing code expands all the arguments.
This means that an out-dir parameter having the same name of an
executable is fully expanded, thus becoming an executable instead of a
directory.

Examples:
   > kcov python echo
   kcov: error: Can't find or open echo

   > kcov python /bin/echo
   kcov: error: Can't open directory /usr/bin/python3.11/

Remove checking if the first argument is an elf file, since the
executable must be after the out-dir argument.

Skip options and don't expand arguments, unless lstat fails.

Issue #414: kcov: error when out-dir is an executable