configuration: improve handling of positional arguments
Currently getopt_long is configured to support mixing options and
positional arguments, but kcov don't use the special "--" argument to
tell getopt_long when to stop parsing the executable arguments.
Instead the implementation tries to scan the arguments to find an
executable, but the implementation is flawed because there are two
required positional arguments, and both may be executable.

In order to handle positional arguments correctly, configure getopt_long
to stop parsing after encountering a positional argument.

This change greatly simplified the code, since the pre scanning can be
removed.  However it may break the workflow of some users.

Add a regression test for issue #414

Issue #414: kcov: error when out-dir is an executable