tests: fix basic.OutDirectoryIsExecutable
Commit 4b4cb9c (Merge pull request #426 from perillo/fix-argv-parsing)
added a regression test, but the test failed due to ptrace errors on
Linux.
The "Operation not permitted" error was caused by kcov trying to trace
the "/bin/usr/python".

The solution is to execute an executable having user permission.
Restore the short-test.py executable, and use it instead of python.
short-test.py always exits with exit status 0.