tests: use subprocess.run to run child process
The current, old API, is not only inconvenient (using a thread and a
timer to implement timeouts) but it is also incorrect since the child
process is not waited.

Use subprocess.run for both `doShell` and `do`.  Note that for the `do`
method, this is a behavior change, since a child process is always
terminated by kill.

Add a default timeout, set to 10 min.

After this change the test `system_mode_can_record_and_report_binary`
fails.  This will be addressed in the next commit.