tests: add testbase.KcovTestCase.doCmd method
The new function is meant to be used when running a child process that
does not run kcov, both directly or indirectly.  The method call usually
returns noKcovRv, instead of rv.

The reason for this change is to make these calls easier to detect, also
avoiding having to set the kcovKcov parameter to False.