tests: make calls to KcovTestCase.do more strict
Update the KcovTestCase.do method signature, so that kcovKcov
is both a positional or keyword parameter, and the timeout is a keyword
only parameter.

I was planning to make kcovKcov a positional only parameter, but the
`bash_linux_only.bash_exit_before_child` test used it as a keyword
parameter.

This change codifies the current usage.