tests: refactor the test runner
The current implementation, using the unittest.main function, is not
working correctly with how kcov tests are loaded.  One annoying problem
is that filtering tests is not possible.

Rewrite the test runner, adding a new libkcov package.

The code from "testbase.py" is moved to libkcov.__init__ (the test API),
and the code from "run-tests" is moved to libkcov.main (the test
runner).

In the test API, remove the configure method and the global variables,
since the configuration paths are now processed by argparse as
positional arguments.  These paths are validated and then used to
initialize each test case, that now has an __init__ method.

In order to make running the tests convenient, add a __main__ entry
point to the libkcov package.

Rename the test names; "file.py" is renamed to "test_file.py", since
this is the preferred naming conventions.

Tests are loaded by a custom test loader, not derived from
unittest.TestCase, since the code is really simple.

Move the "cobertura.py" module to the libkcov package, but without the
main code.  The main code has been moved to "parse_cobertura", so that
it can be used as a tool.

The new implementation make is easy do separate the test API, test
runner, tests and tools.

In order to reduce the number of changes, in each test file libkcov is
imported using as testbase.  It will updated in a separate commit.

Update the "ci-run-tests.sh" script to run the tests with `python -m
linbkcov`.  It is necessary to set the PYTHONPATH environment variable.

The behavior of the new test runner implementation should be the same as
the old one, with the difference that the order of tests is different.
The new order is what users expect.