tests: fix incorrect code in test_compiled.py
The `SignalBase` method of the SignalBase class is never called;
probably it should have be the __init__ method.  Remove it, since
`self.m_self` is always set on subclasses.

In `collect_and_report_only`, use `Exception`, instead of
`BaseException`, since the former does not catch `GeneratorExit`,
`KeyboardInterrupt` and `SystemExit`.

In `system_mode_can_record_and_report_binary`, the test fails as
expected, but the cause of the error is:
    NameError: name 'time' is not defined

Import the `time` module.

Errors reported by `ruff check`.